fit, predict, cv

cv.jl

This unit implements cross-validation procedures for estimating the accuracy and balanced accuracy of machine learning models. It also reports the documentation of the fit and predict functions, as they are common to all models.

Content

struct	description

CVacc	encapsulate the result of cross-validation procedures for estimating accuracy

function	description

fit	fit a model with training data, or create and fit it

predict	preidct labels, probabilities or scoring functions on test data

cvAcc	estimate accuracy of a model by cross-validation

cvSetup	generate indexes for performing cross-validtions

PosDefManifoldML.CVacc — Type.

struct CVacc
    cvType    :: String
    scoring   :: Union{String, Nothing}
    modelType :: Union{String, Nothing}
    cnfs      :: Union{Vector{Matrix{T}}, Nothing} where T<:Real
    avgCnf    :: Union{Matrix{T}, Nothing} where T<:Real
    accs      :: Union{Vector{T}, Nothing} where T<:Real
    avgAcc    :: Union{Real, Nothing}
    stdAcc    :: Union{Real, Nothing}
end

A call to cvAcc results in an instance of this structure. Fields:

.cvTpe is the type of cross-validation technique, given as a string (e.g., "10-kfold")

.scoring is the type of accuracy that is computed, given as a string. This has been passed as argument to cvAcc. Currently accuracy and balanced accuracy are supported.

.modelType is type of the machine learning used for performing the cross-validation, given as a string.

.cnfs is a vector of matrices holding the confusion matrices obtained at each fold of the cross-validation.

.avgCnf is the average confusion matrix across the folds of the cross-validation.

.accs is a vector of real numbers holding the accuracies obtained at each fold of the cross-validation.

.avgAcc is the average accuracy across the folds of the cross-validation.

.stdAcc is the standard deviation of the accuracy across the folds of the cross-validation.

source

StatsBase.fit — Function.

function fit(model :: MDMmodel,
              𝐏Tr   :: ℍVector,
              yTr   :: Vector;
       w        :: Vector = [],
       ✓w       :: Bool  = true,
       meanInit :: Union{ℍVector, Nothing} = nothing,
       tol      :: Real  = 1e-7,
       verbose  :: Bool  = true,
       ⏩       :: Bool  = true)

Fit an MDM machine learning model, with training data 𝐏Tr, of type ℍVector, and corresponding labels yTr, of type IntVector. Return the fitted model.

Labels must be provided using the natural numbers, i.e., 1 for the first class, 2 for the second class, etc.

Fitting an MDM model involves only computing a mean of all the matrices in each class. Those class means are computed according to the metric specified by the MDM constructor.

Optional keyword argument w is a vector of non-negative weights associated with the matrices in 𝐏Tr. This weights are used to compute the mean for each class. See method (3) of the mean function for the meaning of the arguments w, ✓w and ⏩, to which they are passed. Keep in mind that here the weights should sum up to 1 separatedly for each class, which is what is ensured by this function if ✓w is true.

Optional keyword argument tol is the tolerance required for those algorithms that compute the mean iteratively (they are those adopting the Fisher, logde0 or Wasserstein metric). For details on this argument see the functions that are called for computing the means:

Fisher metric: gmean

logdet0 metric: ld0mean

Wasserstein metric: Wasmean.

For those algorithm an initialization can be provided with optional keyword argument meanInit. If provided, this must be a vector of Hermitian matrices of the ℍVector type and must contain as many initializations as classes, in the natural order corresponding to the class labels (see above).

If verbose is true (default), information is printed in the REPL. This option is included to allow repeated calls to this function without crowding the REPL.

See: notation & nomenclature, the ℍVector type.

See also: predict, cvAcc.

Examples

using PosDefManifoldML

# generate some data
PTr, PTe, yTr, yTe=gen2ClassData(10, 30, 40, 60, 80, 0.25)

# create and fit a model:
m=fit(MDM(Fisher), PTr, yTr)

source

function fit(model  :: ENLRmodel,
               𝐏Tr  :: Union{ℍVector, Matrix{Float64}},
               yTr  :: IntVector;
           w        :: Union{Symbol, Tuple, Vector} = [],
           meanISR  :: Union{ℍ, Nothing} = nothing,
		   meanInit :: Union{ℍ, Nothing} = nothing,
           fitType  :: Symbol = :best,
           vecRange :: UnitRange = 𝐏Tr isa ℍVector ? (1:size(𝐏Tr[1], 2)) : (1:size(𝐏Tr, 2)),
           verbose  :: Bool = true,
           ⏩      :: Bool = true,
           # arguments for `GLMNet.glmnet` function
           alpha            :: Real = model.alpha==nothing ? 1.0 : model.alpha,
           weights          :: Vector{Float64} = ones(Float64, length(yTr)),
           intercept        :: Bool = true,
           standardize      :: Bool = alpha≈1.0 ? true : false,
           penalty_factor   :: Vector{Float64} = ones(Float64, _getDim(𝐏Tr, vecRange)),
           constraints      :: Matrix{Float64} = [x for x in (-Inf, Inf), y in 1:_getDim(𝐏Tr, vecRange)],
           offsets          :: Union{Vector{Float64}, Nothing} = nothing,
           dfmax            :: Int = _getDim(𝐏Tr, vecRange),
           pmax             :: Int = min(dfmax*2+20, _getDim(𝐏Tr, vecRange)),
           nlambda          :: Int = 100,
           lambda_min_ratio :: Real = (length(yTr) < _getDim(𝐏Tr, vecRange) ? 1e-2 : 1e-4),
           lambda           :: Vector{Float64} = Float64[],
           tol              :: Real = 1e-7,
           maxit            :: Int = 1000000,
           algorithm        :: Symbol = :newtonraphson,
           # selection method
           λSelMeth :: Symbol = :sd1,
           # arguments for `GLMNet.glmnetcv` function
           nfolds   :: Int = min(10, div(size(yTr, 1), 3)),
           folds    :: Vector{Int} =
           begin
               n, r = divrem(size(yTr, 1), nfolds)
               shuffle!([repeat(1:nfolds, outer=n); 1:r])
           end,
           parallel ::Bool=false)

Create and fit an ENLR machine learning model, with training data 𝐏Tr, of type ℍVector, and corresponding labels yTr, of type IntVector. Return the fitted model.

As for all ML models acting in the tangent space, fitting an ENLR model involves computing a mean of all the matrices in 𝐏Tr, mapping all matrices onto the tangent space after parallel transporting them at the identity matrix and vectorizing them using the vecP operation. Once this is done, the elastic net logistic regression is fitted.

The mean is computed according to the .metric field of the model, with optional weights w. The .metric field of the model is passed to the tsMap function. By default the metric is the Fisher metric. See the examples here below to see how to change metric. See mdm.jl for the available metrics.

Optional keyword arguments

By default, uniform weights will be given to all observations for computing the mean to pass in the tangent space. This is equivalent to passing as argument w=:uniform (or w=:u). You can also pass as argument:

w=:balanced (or simply w=:b). If the two classes are unbalanced, the weights should be inversely proportional to the number of examples for each class, in such a way that each class contributes equally to the computation of the mean. This is equivalent of passing w=tsWeights(yTr), see the tsWeights function for details.

w=v, where v is a user defined vector of non-negative weights for the observations, thus, v must contain the same number of elements as yTr. For example, w=[1.0, 1.0, 2.0, 2.0, ...., 1.0]

w=t, where t is a 2-tuple of real weights, one weight for each class, for example w=(0.5, 1.5). This is equivalent to passing w=tsWeights(yTr; classWeights=collect(t)), see the tsWeights function for details.

If meanISR is passed as argument, the mean is not computed, instead this matrix is the inverse square root (ISR) of the mean used for projecting the matrices in the tangent space (see tsMap). Passed or computed, the inverse square root (ISR) of the mean will be written in the .meanISR field of the created ENLR struct. If meanISRis is not provided and the .metric field of the model is Fisher, logdet0 or Wasserstein, the tolerance of the iterative algorithm used to compute the mean is set to the argument passed as tol (default 1e-7). Also, in this case a particular initialization for those iterative algorithms can be provided as an Hermitian matrix with argument meanInit.

This function also allows to fit a model passing as training data 𝐏Tr directly a matrix of feature vectors, where each feature vector is a row of the matrix. In this case the metric of the ENLR model and argument meanISR are not used. Therefore, the .meanISR field of the created ENLR struct will be set to nothing.

If fitType = :best (default), a cross-validation procedure is run to find the best lambda hyperparameter for the training data. This finds a single model that is written into the .best field of the model that will be created.

If fitType = :path, the regularization path for several values of the lambda hyperparameter if found for the training data. This creates several models, which are written into the .path field of the model that will be created, none of which is optimal, in the cross-validation sense, for the training data.

If fitType = :all, both the above fits are performed and all fields of the model that will be created will be filled in.

If a UnitRange is passed with optional keyword argument vecRange, then if 𝐏Tr is a vector of Hermitianmatrices, the vectorization of those matrices once they are projected onto the tangent space concerns only the rows (or columns) given in the specified range, else if 𝐏Tr is a matrix of feature vectors in the rows, then only the columns of 𝐏Tr given in the specified range will be used.

If verbose is true (default), information is printed in the REPL. This option is included to allow repeated calls to this function without crowding the REPL.

The ⏩ argument (true by default) is passed to the tsMap function for projecting the matrices in 𝐏Tr onto the tangent space using multi-threading.

The remaining optional keyword arguments, are

the arguments passed to the GLMNet.glmnet function for fitting the models. Those are always used.

the λSelMeth argument and the arguments passed to the GLMNet.glmnetcv function for finding the best lambda hyperparamater by cross-validation. Those are used only if fitType = :path or = :all.

Optional keyword arguments for fitting the model(s) using GLMNet

alpha: the hyperparameter in $[0, 1]$ to trade-off an elestic-net model. $α=0$ requests a pure ridge model and $α=1$ a pure lasso model. This defaults to 1.0, which specifies a lasso model, unless the input ENLR model has another value in the alpha field, in which case this value is used. If argument alpha is passed here, it will overwrite the alpha field of the input model.

weights: a vector of weights for each matrix (or feature vectors) of the same size as yTr. It defaults to 1 for all matrices.

intercept: whether to fit an intercept term. The intercept is always unpenalized. Defaults to true.

standardize: whether to standardize predictors so that they are in the same units. Differently from GLMNet.jl, by default this is true for lasso models ($α=1$), false otherwise ($0≤α<1$).

penalty_factor: a vector of length $n(n+1)/2$, where $n$ is the dimension of the original PD matrices on which the model is applied, of penalties for each predictor in the tangent vectors. This defaults to all ones, which weights each predictor equally. To specify that a predictor should be unpenalized, set the corresponding entry to zero.

constraints: an $[n(n+1)/2]$ x $2$ matrix specifying lower bounds (first column) and upper bounds (second column) on each predictor. By default, this is [-Inf Inf] for each predictor (each element of tangent vectors).

dfmax: The maximum number of predictors in the largest model.

pmax: The maximum number of predictors in any model.

nlambda: The number of values of $λ$ along the path to consider.

lambda_min_ratio: The smallest $λ$ value to consider, as a ratio of the value of $λ$ that gives the null model (i.e., the model with only an intercept). If the number of observations exceeds the number of variables, this defaults to 0.0001, otherwise 0.01.

lambda: The $λ$ values to consider for fitting. By default, this is determined from nlambda and lambda_min_ratio.

tol: Convergence criterion. Defaults to 1e-7.

maxit: The maximum number of iterations of the cyclic coordinate descent algorithm. If convergence is not achieved, a warning is returned.

algorithm: the algorithm used to find the regularization path. Possible values are :newtonraphson (default) and :modifiednewtonraphson.

For further informations on those arguments, refer to the resources on the GLMNet package 🎓.

Optional Keyword arguments for finding the best model by cv

λSelMeth = :sd1 (default), the best model is defined as the one allowing the highest cvλ.meanloss within one standard deviation of the minimum, otherwise it is defined as the one allowing the minimum cvλ.meanloss. Note that in selecting a model, the model with only the intercept term, if it exists, is ignored. See ENLRmodel for a description of the .cvλ field of the model structure.

Arguments nfolds, folds and parallel are passed to the GLMNet.glmnetcv function. Please refer to the resources on GLMNet for details 🎓.

See: notation & nomenclature, the ℍVector type.

See also: predict, cvAcc.

Examples

using PosDefManifoldML

# generate some data
PTr, PTe, yTr, yTe=gen2ClassData(10, 30, 40, 60, 80, 0.1)

# Fit an ENLR lasso model and find the best model by cross-validation:
m=fit(ENLR(), PTr, yTr)

# ... balancing the weights for tangent space mapping
m=fit(ENLR(), PTr, yTr; w=tsWeights(yTr))

# ... using the log-Eucidean metric for tangent space projection
m=fit(ENLR(logEuclidean), PTr, yTr)

# Fit an ENLR ridge model and find the best model by cv:
m=fit(ENLR(Fisher), PTr, yTr; alpha=0)

# Fit an ENLR elastic-net model (α=0.9) and find the best model by cv:
m=fit(ENLR(Fisher), PTr, yTr; alpha=0.9)

# Fit an ENLR lasso model and its regularization path:
m=fit(ENLR(), PTr, yTr; fitType=:path)

# Fit an ENLR lasso model, its regularization path
# and the best model found by cv:
m=fit(ENLR(), PTr, yTr; fitType=:all)

source

GLMNet.predict — Function.

function predict(model  :: MDMmodel,
                 𝐏Te    :: ℍVector,
                 what   :: Symbol = :labels;
               verbose :: Bool = true,
               ⏩     :: Bool = true)

Given an MDM model trained (fitted) on $z$ classes and a testing set of $k$ positive definite matrices 𝐏Te of type ℍVector,

if what is :labels or :l (default), return the predicted class labels for each matrix in 𝐏Te, as an IntVector. For MDM models, the predicted class 'label' of an unlabeled matrix is the serial number of the class whose mean is the closest to the matrix (minimum distance to mean). The labels are '1' for class 1, '2' for class 2, etc;

if what is :probabilities or :p, return the predicted probabilities for each matrix in 𝐏Te to belong to a all classes, as a $k$-vector of $z$ vectors holding reals in $[0, 1]$ (probabilities). The 'probabilities' are obtained passing to a softmax function minus the squared distances of each unlabeled matrix to all class means;

if what is :f or :functions, return the output function of the model. The ratio of the squared distance to all classes to their geometric mean gives the 'functions'.

If verbose is true (default), information is printed in the REPL. This option is included to allow repeated calls to this function without crowding the REPL.

It f ⏩ is true (default), the computation of distances is multi-threaded.

See: notation & nomenclature, the ℍVector type.

See also: fit, cvAcc, predictErr.

Examples

using PosDefManifoldML

# generate some data
PTr, PTe, yTr, yTe=gen2ClassData(10, 30, 40, 60, 80)

# craete and fit an MDM model
m=fit(MDM(Fisher), PTr, yTr)

# predict labels
yPred=predict(m, PTe, :l)

# prediction error
predErr=predictErr(yTe, yPred)

# predict probabilities
predict(m, PTe, :p)

# output functions
predict(m, PTe, :f)

source

function predict(model   :: ENLRmodel,
                 𝐏Te     :: Union{ℍVector, Matrix{Float64}},
                 what    :: Symbol = :labels,
                 fitType :: Symbol = :best,
                 onWhich :: Int    = Int(fitType==:best);
			transfer   :: Union{ℍ, Nothing} = nothing,
            vecRange   :: UnitRange = 𝐏Te isa ℍVector ? (1:size(𝐏Te[1], 2)) : (1:size(𝐏Te, 2)),
            checks     :: Bool = true,
            verbose    :: Bool = true,
            ⏩        :: Bool = true)

Given an ENLR model trained (fitted) on 2 classes and a testing set of $k$ positive definite matrices 𝐏Te of type ℍVector,

if what is :labels or :l (default), return the predicted class labels for each matrix in 𝐏Te, as an IntVector. Those labels are '1' for class 1 and '2' for class 2;

if what is :probabilities or :p, return the predicted probabilities for each matrix in 𝐏Te to belong to a all classes, as a $k$-vector of $z$ vectors holding reals in $[0, 1]$ (probabilities). The 'probabilities' are obtained passing to a softmax function the output of the ENLR model and zero;

if what is :f or :functions, return the output function of the model, which is the raw output of the ENLR model.

If fitType = :best (default), the best model that has been found by cross-validation is used for prediction.

If fitType = :path,

if onWhich is a valid serial number for a model in the model.path, this model is used for prediction,

if onWhich is zero, all model in the model.path will be used for predictions, thus the output will be multiplied by the number of models in model.path.

Argumet onWhich has no effect if fitType = :best.

Nota Bene

By default, the fit function fits only the best model. If you want to use the fitType = :path option you need to invoke the fit function with optional keyword argument fitType=:path or fitType=:all. See the fit function for details.

Option keyword argument transfer can be used to specify the principal inverse square root (ISR) of a new mean to be used as base point for projecting the matrices in 𝐏Te onto the tangent space. By default transfer is equal to nothing, implying that the base point will be the mean used to fit the model. Passing a new mean ISR allows the adaptation first described in Barachant et al.(2013). Typically transfer is the ISR of the mean of the matrices in 𝐏Te or of a subset of them. Notice that this actually performs transfer learning by parallel transporting both the training and test data to the identity matrix as defined in Zanini et al.(2018) and taken up in Rodrigues et al.(2019)🎓.

Optional keyword argument vecRange has the same meaning as in the fit function. In general, if you have used it to fit the model youwill use the same here.

If checks is true (default), checks on the validity of the arguments are performed. This can be set to false to spped up computations.

If verbose is true (default), information is printed in the REPL. This option is included to allow repeated calls to this function without crowding the REPL.

If ⏩ = true (default) and 𝐏Te is an ℍVector type, the projection onto the tangent space is multi-threaded.

See: notation & nomenclature, the ℍVector type.

See also: fit, cvAcc, predictErr.

Examples

using PosDefManifoldML

# generate some data
PTr, PTe, yTr, yTe=gen2ClassData(10, 30, 40, 60, 80)

# fit an ENLR lasso model and find the best model by cv
m=fit(ENLR(Fisher), PTr, yTr)

# predict labels from the best model
yPred=predict(m, PTe, :l)
# prediction error
predErr=predictErr(yTe, yPred)

# predict probabilities from the best model
predict(m, PTe, :p)

# output functions from the best model
predict(m, PTe, :f)

# fit a regularization path for an ENLR lasso model
m=fit(ENLR(Fisher), PTr, yTr; fitType=:path)

# predict labels using a specific model
yPred=predict(m, PTe, :l, :path, 10)

# predict labels for all models
yPred=predict(m, PTe, :l, :path, 0)
# prediction error for all models
predErr=[predictErr(yTe, yPred[:, i]) for i=1:size(yPred, 2)]

# predict probabilities from a specific model
predict(m, PTe, :p, :path, 12)

# predict probabilities from all models
predict(m, PTe, :p, :path, 0)

# output functions from specific model
predict(m, PTe, :f, :path, 3)

# output functions for all models
predict(m, PTe, :f, :path, 0)

source

PosDefManifoldML.cvAcc — Function.

function cvAcc(model   :: MLmodel,
               𝐏Tr     :: ℍVector,
               yTr     :: IntVector;
           tol       :: Real      = 0.,
           nFolds    :: Int       = min(10, length(yTr)÷3),
           scoring   :: Symbol    = :b,
           shuffle   :: Bool      = false,
           vecRange  :: UnitRange = 𝐏Tr isa ℍVector ? (1:size(𝐏Tr[1], 2)) : (1:size(𝐏Tr, 2)),
           verbose   :: Bool      = true,
           outModels :: Bool      = false,
           fitArgs...)

Cross-validation accuracy for a machine learning model: given an ℍVector 𝐏Tr holding $k$ Hermitian matrices, an IntVector yTr holding the $k$ labels for these matrices and the number of folds nFolds, return a CVacc structure.

optional keyword arguments

The Fisher, logdet0 and Wasserstein metric of the machine learning models require an iterative algorithm for computing means in the manifold of positive definite matrices. Argument tol is the tolerance for convergence of these iterative algorithms. In order to speed up computations, set tol to something between 10e-6 (still a fairly good convergence) and 10e-3 (coarse convergence, hence possible drop of classification accuracy, but many less iterations required).

nFolds by default is set to the minimum between 10 and the number of observation ÷ 3 (integer division).

If scoring=:b (default) the balanced accuracy is computed. Any other value will make the function returning the regular accuracy. Balanced accuracy is to be preferred for unbalanced classes. For balanced classes the balanced accuracy reduces to the regular accuracy, therefore there is no point in using regular accuracy if not to avoid a few unnecessary computations when the class are balanced.

For the meaning of the shuffle argument (false by default), see function cvSetup, to which this argument is passed.

Argument vecRange has an effect only for machine learning models in the tangent space. For its meaning see function fit and predict, to which it is passed for each fold.

If verbose is true (default), information is printed in the REPL. This option is included to allow repeated calls to this function without crowding the REPL.

if outModelsis true return a 2-tuple holding a [CVacc](@ref) structure and anFolds-vector of the model fitted for each fold, otherwise (default), return only a [CVacc`](@ref) structure.

fitArgs are optional keyword arguments that are passed to the fit function called for each fold of the cross-validation. For each machine learning model, all optional keyword arguments of their fit method are elegible to be passed here, however, the arguments listed in the following table for each model should not be passed. Note that if they are passed, they will be disabled:

MDM/MDMF	ENLR

verbose	verbose

⏩	⏩

meanISR

fitType

offsets

lambda

folds

See: notation & nomenclature, the ℍVector type.

See also: fit, predict.

Examples

using PosDefManifoldML

# generate some data
PTr, PTe, yTr, yTe=gen2ClassData(10, 30, 40, 60, 80)

# perform 10-fold cross-validation using the minimum distance to mean classifier
cv=cvAcc(MDM(Fisher), PTr, yTr)

# ...using the lasso logistic regression classifier
cv=cvAcc(ENLR(Fisher), PTr, yTr)

# perform 8-fold cross-validation instead
cv=cvAcc(ENLR(Fisher), PTr, yTr; nFolds=8)

# ...using the elastic-net logistic regression (α=0.9) classifier
cv=cvAcc(ENLR(Fisher), PTr, yTr; nFolds=8, alpha=0.9)

# ...and standardizing the predictors
cv=cvAcc(ENLR(Fisher), PTr, yTr; nFolds=8, alpha=0.9, standardize=true)

# perform another cross-validation shuffling the folds
cv=cvAcc(MDM(Fisher), PTr, yTr; nFolds=8, shuffle=true)

source

PosDefManifoldML.cvSetup — Function.

function cvSetup(k       :: Int,
                 nCV     :: Int;
                 shuffle :: Bool = false)

Given k elements and a parameter nCV, a nCV-fold cross-validation is obtained defining $nCV$ permutations of $k$ elements in $nTest=k÷nCV$ (integer division) elements for the test and $k-nTest$ elements for the training, in such a way that each element is represented in only one permutation.

Said differently, given a length k and the number of desired cross-validations nCV, this function generates indices from the sequence of natural numbers $1,..,k$ to obtain all nCV-fold cross-validation sets. Specifically, it generates $nCV$ vectors of indices for generating test sets and $nCV$ vectors of indices for geerating training sets.

If optional keyword argument shuffle is true, the sequence of natural numbers $1,..,k$ is shuffled before running the function, thus in this case two successive runs of this function will give different cross-validation sets, hence different accuracy scores. By default shuffle is false, so as to allow exactly the same result in successive runs. Note that no random initialization for the shuffling is provided, so as to allow the replication of the same random sequences starting again the random generation from scratch.

This function is used in cvAcc. It constitutes the fundamental basis to implement customized cross-validation procedures.

Return the 2-tuple with:

A vector of nCV vectors holding the indices for the training sets,

A vector of nCV vectors holding the indices for the corresponding test sets.

Examples

using PosDefManifoldML

cvSetup(10, 2)
# return:
# (Array{Int64,1}[[6, 7, 8, 9, 10], [1, 2, 3, 4, 5]],
#  Array{Int64,1}[[1, 2, 3, 4, 5], [6, 7, 8, 9, 10]])

cvSetup(10, 2, shuffle=true)
# return:
# (Array{Int64,1}[[5, 4, 6, 1, 9], [3, 7, 8, 2, 10]],
#  Array{Int64,1}[[3, 7, 8, 2, 10], [5, 4, 6, 1, 9]])

cvSetup(10, 3)
# return:
# (Array{Int64,1}[[4, 5, 6, 7, 8, 9, 10], [1, 2, 3, 7, 8, 9, 10], [1, 2, 3, 4, 5, 6]],
#  Array{Int64,1}[[1, 2, 3], [4, 5, 6], [7, 8, 9, 10]])

source