fit, predict, cv

cv.jl

This unit implements cross-validation procedures for estimating the accuracy and balanced accuracy of machine learning models. It also reports the documentation of the fit and predict functions, as they are common to all models.

Content

struct	description

CVacc	encapsulate the result of cross-validation procedures for estimating accuracy

function	description

fit	fit a model with training data, or create and fit it

predict	preidct labels, probabilities or scoring functions on test data

cvAcc	estimate accuracy of a model by cross-validation

cvSetup	generate indexes for performing cross-validtions

PosDefManifoldML.CVacc — Type.

struct CVacc
    cvType  :: String
    scoring :: Union{String, Nothing}
    model   :: Union{MLmodel, Nothing}
    cnfs    :: Union{Vector{Matrix{T}}, Nothing} where T<:Real
    avgCnf  :: Union{Matrix{T}, Nothing} where T<:Real
    accs    :: Union{Vector{T}, Nothing} where T<:Real
    avgAcc  :: Union{Real, Nothing}
    stdAcc  :: Union{Real, Nothing}
end

A call to cvAcc results in an instance of this structure. Fields:

.cvTpe is the type of cross-validation technique, given as a string (e.g., "10-kfold")

.scoring is the type of accuracy that is computed, given as a string. This has been passed as argument to cvAcc. Currently accuracy and balanced accuracy are supported.

.model is the machine learning model that has been passed as argument to cvAcc, e.g., an MDMmodel or an ENLRmodel.

.cnfs is a vector of matrices holding the confusion matrices obtained at each fold of the cross-validation.

.avgCnf is the average confusion matrix across the folds of the cross-validation.

.accs is a vector of real numbers holding the accuracies obtained at each fold of the cross-validation.

.avgAcc is the average accuracy across the folds of the cross-validation.

.stdAcc is the standard deviation of the accuracy across the folds of the cross-validation.

source

StatsBase.fit — Function.

function fit(model :: MDMmodel,
              𝐏Tr   :: ℍVector,
              yTr   :: Vector;
           w       :: Vector = [],
           ✓w      :: Bool  = true,
           verbose :: Bool  = true,
           ⏩      :: Bool  = true)

Fit an MDM machine learning model, with training data 𝐏Tr, of type ℍVector, and corresponding labels yTr, of type IntVector. Return the fitted model.

Fitting an MDM model involves only computing a mean of all the matrices in each class. Those class means are computed according to the metric specified by the MDM constructor.

See method (3) of the mean function for the meaning of the optional keyword arguments w, ✓w and ⏩, to which they are passed. Keep in mind that here the weights should sum up to 1 separatedly for each class, which is what is ensured if ✓w is true.

If verbose is true (default), information is printed in the REPL. This option is included to allow repeated calls to this function without crowding the REPL.

See: notation & nomenclature, the ℍVector type.

See also: predict, cvAcc.

Examples

using PosDefManifoldML

# generate some data
PTr, PTe, yTr, yTe=gen2ClassData(10, 30, 40, 60, 80, 0.25)

# create and fit a model:
m=fit(MDM(Fisher), PTr, yTr)

source

function fit(model :: ENLRmodel,
              𝐏Tr   :: Union{ℍVector, Matrix{Float64}},
              yTr   :: Vector;
           w       :: Vector            = [],
           meanISR :: Union{ℍ, Nothing} = nothing,
           verbose :: Bool              = true,
           ⏩     :: Bool               = true,
           # (GLMNet) fitArgs
           alpha            :: Real = model.alpha < 1.0 ? model.alpha : 1.0,
           intercept        :: Bool = true,
           standardize      :: Bool = alpha≈1.0 ? true : false,
           penalty_factor   :: Vector{Float64} = ones(_triNum(𝐏Tr[1])),
           constraints      :: Matrix{Float64} = [x for x in (-Inf, Inf), y in 1:_triNum(𝐏Tr[1])],
           offsets          :: Union{Vector{Float64}, Nothing} = nothing,
           dfmax            :: Int = _triNum(𝐏Tr[1]),
           pmax             :: Int = min(dfmax*2+20, _triNum(𝐏Tr[1])),
           nlambda          :: Int = 100,
           lambda_min_ratio :: Real = (length(yTr) < _triNum(𝐏Tr[1]) ? 1e-2 : 1e-4),
           lambda           :: Vector{Float64} = Float64[],
           tol              :: Real = 1e-7,
           maxit            :: Int = 1000000,
           algorithm        :: Symbol = :newtonraphson)

Fit an ENLR machine learning model, with training data 𝐏Tr, of type ℍVector, and corresponding labels yTr, of type IntVector. Return the fitted model.

Fitting an ENLR model involves computing a mean of all the matrices in 𝐏Tr, mapping all matrices onto the tangent space after parallel transporting them at the identity matrix and vectorizing them using the vecP operation. Once this is done the elastic net logistic regression is fitted.

The mean is computed according to the .metric field of the model, with optional weights w. If weights are used, they should be inversely proportional to the number of examples for each class, in such a way that each class contributes equally to the computation of the mean. This argument is passed to the tsMap function.

If meanISR is passed as argument, the mean is not computed, instead this matrix is the inverse square root (ISR) of the mean used for projecting the matrices in the tangent space (see tsMap).

If verbose is true (default), information is printed in the REPL. This option is included to allow repeated calls to this function without crowding the REPL.

The ⏩ argument (true by default) is passed to the tsMap function for projecting the matrices in 𝐏Tr onto the tangent space.

The remaining optional arguments, referred to as fitArgs, are passed to the GLMNet.glmnet function to fit the elastic-net model.

weights: a vector of weights for each matrix of the same size as yTr. Argument w is passed, defaulting to 1 for all matrices.

alpha: the hyperparameter in $[0, 1]$ to trade-off an elestic-net model. $α=0$ requests a pure ridge model and $α=1$ a pure lasso model. This defaults to 1.0, which specifies a lasso model.

intercept: whether to fit an intercept term. The intercept is always unpenalized. Defaults to true.

standardize: whether to standardize predictors so that they are in the same units. Differently from GLMNet.jl, by default this is true for lasso models ($α=1$), false otherwise ($0≤α<1$).

penalty_factor: a vector of length $n(n+1)/2$, where $n$ is the dimension of the original PD matrices on which the model is applied, of penalties for each predictor in the tangent vectors. This defaults to all ones, which weights each predictor equally. To specify that a predictor should be unpenalized, set the corresponding entry to zero.

constraints: an $n(n+1)/2$ x $2$ matrix specifying lower bounds (first column) and upper bounds (second column) on each predictor. By default, this is [-Inf Inf] for each predictor (each element of tangent vectors).

dfmax: The maximum number of predictors in the largest model.

pmax: The maximum number of predictors in any model.

nlambda: The number of values of $λ$ along the path to consider.

lambda_min_ratio: The smallest $λ$ value to consider, as a ratio of the value of $λ$ that gives the null model (i.e., the model with only an intercept). If the number of observations exceeds the number of variables, this defaults to 0.0001, otherwise 0.01.

lambda: The $λ$ values to consider. By default, this is determined from nlambda and lambda_min_ratio.

tol: Convergence criterion. Defaults to 1e-7.

maxit: The maximum number of iterations of the cyclic coordinate descent algorithm. If convergence is not achieved, a warning is returned.

algorithm: the algorithm used to find the path.

For further informations on the fitArgs arguments, refer to the resources on the GLMNet package 🎓.

See: notation & nomenclature, the ℍVector type.

See also: predict, cvAcc.

Examples

using PosDefManifoldML

# generate some data
PTr, PTe, yTr, yTe=gen2ClassData(10, 30, 40, 60, 80, 0.25)

# create and fit an ENLR lasso model:
m=fit(ENLR(Fisher), PTr, yTr)

# create and fit an ENLR ridge model:
m=fit(ENLR(Fisher), PTr, yTr; alpha=0)

# create and fit an ENLR elastic net model with alpha=0.9:
m=fit(ENLR(Fisher), PTr, yTr; alpha=0.9)

source

GLMNet.predict — Function.

function predict(model  :: MDMmodel,
                 𝐏Te    :: ℍVector,
                 what   :: Symbol = :labels;
               verbose :: Bool = true,
               ⏩     :: Bool = true)

Given an MDM model trained (fitted) on $z$ classes and a testing set of $k$ positive definite matrices 𝐏Te of type ℍVector,

if what is :labels or :l (default), return the predicted class labels for each matrix in 𝐏Te as an IntVector;

if what is :probabilities or :p, return the predicted probabilities for each matrix in 𝐏Te to belong to a all classes, as a $k$-vector of $z$ vectors holding reals in $[0, 1]$ (probabilities).

if what is :f or :functions, return the output function of the model (see below).

If verbose is true (default), information is printed in the REPL. This option is included to allow repeated calls to this function without crowding the REPL.

It f ⏩ is true (default), the computation of distances is multi-threaded.

For MDM models, the predicted class 'label' of an unlabeled matrix is the serial number of the class whose mean is the closest to the matrix (minimum distance to mean). The labels are '1' for class 1, '2' for class 2, etc.

The 'probabilities' are obtained passing to a softmax function minus the squared distances of each unlabeled matrix to all class means.

The ratio of these squared distance to their geometric mean gives the 'functions'.

See: notation & nomenclature, the ℍVector type.

See also: fit, cvAcc, predictErr.

Examples

using PosDefManifoldML

# generate some data
PTr, PTe, yTr, yTe=gen2ClassData(10, 30, 40, 60, 80)

# craete and fit an MDM model
m=fit(MDM(Fisher), PTr, yTr)

# predict labels
yPred=predict(m, PTe, :l)

# prediction error
predErr=predictErr(yTe, yPred)

# predict probabilities
predict(m, PTe, :p)

# output functions
predict(m, PTe, :f)

source

function predict(model  :: ENLRmodel,
                 𝐏Te    :: Union{ℍVector, Matrix{Float64}},
                 what   :: Symbol = :labels,
                 which  :: Union{Int, Nothing} = model.bestModel;
            verbose :: Bool = true,
            ⏩     :: Bool = true)

Given an ENLR model trained (fitted) on 2 classes and a testing set of $k$ positive definite matrices 𝐏Te of type ℍVector,

if what is :labels or :l (default), return the predicted class labels for each matrix in 𝐏Te as an IntVector;

if what is :probabilities or :p, return the predicted probabilities for each matrix in 𝐏Te to belong to a all classes, as a $k$-vector of $z$ vectors holding reals in $[0, 1]$ (probabilities).

if what is :f or :functions, return the output function of the model (see below).

If verbose is true (default), information is printed in the REPL. This option is included to allow repeated calls to this function without crowding the REPL.

The predicted class 'label' are '1' for class 1 and '2' for class 2.

The 'probabilities' are obtained passing to a softmax function the output of the ENLR model and zero.

The 'functions' are the raw output of the ENLR model.

If the cvLambda! function has been previously called on the model, the best model for the model has been selected and by default this model will be used for prediction, otherwise

if which is a valid serial number for a model in the model.path, this model is used for prediction,

if which is zero, all model in the model.path will be used for predictions, thus the output will be multiplied by the number of models in model.path.

See: notation & nomenclature, the ℍVector type.

See also: fit, cvAcc, predictErr.

Examples

using PosDefManifoldML

# generate some data
PTr, PTe, yTr, yTe=gen2ClassData(10, 30, 40, 60, 80)

# craete and fit an ENLR lasso model
m=fit(ENLR(Fisher), PTr, yTr)

# estimate the best model by cross-validation
cvLambda!(m, PTr, yTr)

# predict labels using the best model
yPred=predict(m, PTe, :l)
# prediction error
predErr=predictErr(yTe, yPred)

# predict labels using a specific model
yPred=predict(m, PTe, :l, 10)

# predict labels for all models
yPred=predict(m, PTe, :l, 0)
# prediction error for all models
predErr=[predictErr(yTe, yPred[:, i]) for i=1:size(yPred, 2)]

# predict probabilities using best model
predict(m, PTe, :p)

# predict probabilities using a specific model
predict(m, PTe, :p, 12)

# predict probabilities for all models
predict(m, PTe, :p, 0)

# output functions using best model
predict(m, PTe, :f)

# output functions using a specific model
predict(m, PTe, :f, 3)

# output functions for all models
predict(m, PTe, :f, 0)

source

PosDefManifoldML.cvAcc — Function.

function cvAcc(model   :: MLmodel,
               𝐏Tr     :: ℍVector,
               yTr     :: IntVector,
               nCV     :: Int;
           scoring   :: Symbol = :b,
           shuffle   :: Bool   = false,
           verbose   :: Bool   = true,
           fitArgs...)

Cross-validation accuracy for a machine learning model: given an ℍVector 𝐏Tr holding $k$ Hermitian matrices, an IntVector yTr holding the $k$ labels for these matrices and the number of cross-validations nCV, return a CVacc structure.

If scoring=:b (default) the balanced accuracy is computed. Any other value will make the function returning the regular accuracy. Balanced accuracy is to be preferred for unbalanced classes. In any case, for balanced classes the balanced accuracy reduces to the regular accuracy, therefore there is no point in using regular accuracy.

See: notation & nomenclature, the ℍVector type.

See also: fit, predict.

Examples

using PosDefManifoldML

# generate some data
PTr, PTe, yTr, yTe=gen2ClassData(10, 30, 40, 60, 80)

# perform cross-validation using the minimum distance to mean classifier
cv=cvAcc(MDM(Fisher), PTr, yTr, 10)

# ...using the lasso logistic regression classifier
cv=cvAcc(ENLR(Fisher), PTr, yTr, 10)

# ...using the elastic-net logistic regression (α=0.9) classifier
cv=cvAcc(ENLR(Fisher), PTr, yTr, 10; alpha=0.9)

# perform another cross-validation
cv=cvAcc(MDM(Fisher), PTr, yTr, 10; shuffle=true)

source

PosDefManifoldML.cvSetup — Function.

function cvSetup(k       :: Int,
                 nCV     :: Int;
                 shuffle :: Bool = false)

Given k elements and a parameter nCV, a nCV-fold cross-validation is obtained defining $nCV$ permutations of $k$ elements in $nTest=k÷nCV$ (integer division) elements for the test and $k-nTest$ elements for the training, in such a way that each element is represented in only one permutation.

Said differently, given a length k and the number of desired cross-validations nCV, this function generates indices from the sequence of natural numbers $1,..,k$ to obtain all nCV-fold cross-validation sets. Specifically, it generates $nCV$ vectors of indices for generating test sets and $nCV$ vectors of indices for geerating training sets.

If optional keyword argument shuffle is true, the sequence of natural numbers $1,..,k$ is shuffled before running the function, thus in this case two successive runs of this function will give different cross-validation sets, hence different accuracy scores. By default shuffle is false, so as to allow exactly the same result in successive runs. Note that no random initialization for the shuffling is provided, so as to allow the replication of the same random sequences starting again the random generation from scratch.

This function is used in cvAcc. It constitutes the fundamental basis to implement customized cross-validation procedures.

Return the 2-tuple with:

A vector of nCV vectors holding the indices for the training sets,

A vector of nCV vectors holding the indices for the corresponding test sets.

Examples

using PosDefManifoldML

cvSetup(10, 2)
# return:
# (Array{Int64,1}[[6, 7, 8, 9, 10], [1, 2, 3, 4, 5]],
#  Array{Int64,1}[[1, 2, 3, 4, 5], [6, 7, 8, 9, 10]])

cvSetup(10, 2, shuffle=true)
# return:
# (Array{Int64,1}[[5, 4, 6, 1, 9], [3, 7, 8, 2, 10]],
#  Array{Int64,1}[[3, 7, 8, 2, 10], [5, 4, 6, 1, 9]])

cvSetup(10, 3)
# return:
# (Array{Int64,1}[[4, 5, 6, 7, 8, 9, 10], [1, 2, 3, 7, 8, 9, 10], [1, 2, 3, 4, 5, 6]],
#  Array{Int64,1}[[1, 2, 3], [4, 5, 6], [7, 8, 9, 10]])

source