Elastic-Net Logistic Regression

enlr.jl

This unit implements the elastic net logistic regression (ENLR) machine learning model on the tangent space for symmetric positive definite (SDP) matrices, i.e., real PD matrices. This model features two hyperparameters: a user-defined alpha hyperparameter, in range $[0, 1]$, where $α=0$ allows a pure Ridge LR model and $α=1$ a pure lasso LR model and the lambda hyperparameter. When the model is fitted, several models for several values of lambda are created. The 'best' model can be estimated by cross-validation using the cvLambda! function.

The lasso model (default) has enjoyed popularity in the field of brain-computer interaces due to the winning score obtained in six international data classification competitions.

The ENLR model is implemented using the Julia package GLMNet.jl. See 🎓 for resources on GLMNet and learn how to use purposefully this model.

The fit and predict functions for the ENRL models are reported in the cv.jl unit, since those are shared by all machine learning models. Here it is reported the ENLRmodel abstract type, the ENLR structure and the cvLambda! function, which allows to estimate the best model using cross-validation.

PosDefManifoldML.ENLRmodel — Type.

abstract type ENLRmodel<:TSmodel end

Abstract type for Elastic Net Logistic Rgression (ENLR) machine learning models. See MLmodel.

source

PosDefManifoldML.ENLR — Type.

mutable struct ENLR <: ENLRmodel
    metric      :: Metric = Fisher;
    alpha       :: Real
    intercept   :: Bool
    standardize :: Bool
    meanISR     :: Union{ℍVector, Nothing}
    featDim     :: Int
    path        :: GLMNet.GLMNetPath
    cvλ         :: GLMNet.GLMNetCrossValidation
    bestModel   :: Int
    bestλ       :: Real
end

ENLR machine learning models are incapsulated in this mutable structure. Fields:

.metric, of type Metric, is to be specified by the user. It is the metric that will be adopted to compute the mean used as base-point for tangent space projection.

All other fields do not correspond to arguments passed upon creation of the model. Instead, they are filled later by the fit and cvLambda! functions:

.alpha is the hyperparameter in $[0, 1]$ trading-off the elestic-net model. $α=0$ requests a pure ridge model and $α=1$ a pure lasso model. This is passed as parameter to the fit function, defaulting therein to $α=1$.

.intercept is true (default) if the logistic regression model has an intercept term.

.standardize. If true, predictors are standardized so that they are in the same units. By default is true for lasso models ($α=1$), false otherwise ($0≤α<1$).

.meanISR is optionally passed to the fit function. By default it is computed thereby.

.featDim is the length of the vectorized tangent vectors. This is given by $n(n+1)/2$, where $n$ is the dimension of the original PD matrices on which the model is applied once they are mapped onto the tangent space.

.path is the following structure created using the GLMNet.jl package when the fit function is invoked:

struct GLMNetPath{F<:Distribution}
    family::F                        # Binomial()
    a0::Vector{Float64}              # intercept values for each solution
    betas::CompressedPredictorMatrix # coefficient values for each solution
    null_dev::Float64                # Null deviance of the model
    dev_ratio::Vector{Float64}       # R^2 values for each solution
    lambda::Vector{Float64}          # lambda values for each solution
    npasses::Int                     # actual number of passes over the
                                     # data for all lamda values
end

.cvλ is the following structure created using the GLMNet.jl package when the cvLambda! function is invoked.

struct GLMNetCrossValidation
    path::GLMNetPath            # the cv path
    nfolds::Int                 # the number of folds for the cv
    lambda::Vector{Float64}     # lambda values for each solution
    meanloss::Vector{Float64}   # mean loss for each solution
    stdloss::Vector{Float64}    # standard deviation of the mean losses
end

.bestModel is the serial number of the model in cvλ.path allowing the minimum cvλ.meanloss, or the highest cvλ.meanloss within one standard deviation of the minimum, as specified by the user when invoking the cvLambda! function.

.bestλ is the value of the lambda hyperparameter corresponding to the bestModel.

Examples:

using PosDefManifoldML

# create an empty model
m = ENLR(Fisher)

# since the Fisher metric is the default metric,
# this is equivalent to
m = ENLR()

Note that in general you need to invoke these constructors only when an MDM model is needed as an argument to a function, otherwise you will create and fit an ENLR model using the fit function.

source

PosDefManifoldML.cvLambda! — Function.

function cvLambda!(model :: ENLRmodel,
                   𝐏Tr   :: Union{ℍVector, Matrix{Float64}},
                   yTr   :: Vector;
             w        :: Vector = [],
             λSelMeth :: Symbol = :sd1,
             verbose  :: Bool   = true,
             ⏩      :: Bool   = true,
             # arguments of GLMNet
             nfolds  :: Int = min(10, div(size(yTr, 1), 3)),
             folds   :: Vector{Int} =
             begin
                 n, r = divrem(size(yTr, 1), nfolds)
                 shuffle!([repeat(1:nfolds, outer=n); 1:r])
             end,
             parallel::Bool=false,
             fitArgs...)

Run a cross-validation to find the bestModel, using training data 𝐏Tr, of type ℍVector, and corresponding labels yTr, of type IntVector. Return the fitted model filling the model.cvλ field.

Nota bene

The model must have been fitted calling the fit function before calling this function.

The optionl w weights are passed to the tsMap function, as in the fit function. By defaults all weights are equal to 1, but here too they should be inversely proportional to the number of examples for each class, in such a way that each class contributes equally to the computation of the mean for projecting the matrices onto the tangent space.

If λSelMeth = :sd1 (default), the best model is the one allowing the highest cvλ.meanloss within one standard deviation of the minimum, otherwise it is the one allowing the minimum cvλ.meanloss. Note that in selecting a model, the model with only the intercept term, which is the first model if the fit function has been run with argument intercept=true (default), is ignored.

If verbose is true (default), information is printed in the REPL. This option is included to allow repeated calls to this function without crowding the REPL.

The ⏩ argument (true by default) is passed to the tsMap function for projecting the matrices in 𝐏Tr onto the tangent space.

Arguments nfolds, folds and parallel are passed to the GLMNet.glmnetcv function. Please refer to the resources on GLMNet for details 🎓.

Arguments fitArgs... may be any set of fitArgs arguments to be passed to the GLMNet.glmnet function. See the documentation of the fit function.

Examples

using PosDefManifoldML

# generate some data
PTr, PTe, yTr, yTe=gen2ClassData(10, 30, 40, 60, 80, 0.25)

# create and fit an ENLR lasso model:
m=fit(ENLR(Fisher), PTr, yTr)

# estimate the best model by cross-validation
cvLambda!(m, PTr, yTr)

source