Tools

tools.jl

This unit implements tools that are useful for building Riemannian and Euclidean machine learning classifiers.

Content

function	description

tsMap	project data on a tangent space to apply Euclidean ML models therein

gen2ClassData	generate 2-class positive definite matrix data for testing Riemannian ML models

predictErr	prediction error given a vector of true labels and a vector of predicted labels

PosDefManifoldML.tsMap — Function.

function tsMap(	metric :: Metric,
		𝐏 :: ℍVector;
		w :: Vector = [],
		✓w :: Bool = true,
		⏩ :: Bool = true,
		meanISR :: Union{ℍ, Nothing} = nothing,
		transpose :: Bool = true)

The tangent space mapping of matrices $P_i$, $i=1...k$ with geometric mean $G$, once those points have been parallel transported to the identity matrix, is given by:

$S_i=\textrm{log}(G^{-1/2} P_i G^{-1/2})$.

Given a vector of $k$ Hermitian matrices 𝐏, return a matrix $X$ with such tangent vectors of the matrices in 𝐏 vectorized as per the vecP operation.

The mean $G$ of the matrices in 𝐏 is found according to the specified metric, of type Metric. A natural choice is the Fisher metric.

A set of $k$ optional non-negative weights w can be provided for computing instead the weighted mean $G$. If w is non-empty and optional keyword argument ✓w is true (default), the weights are normalized so as to sum up to 1, otherwise they are used as they are passed and should be already normalized. This option is provided to allow calling this function repeatedly without normalizing the same weights vector each time.

If an Hermitian matrix is provided as optional keyword argument meanISR, then the mean $G$ is not computed, intead this matrix is used directly in the formula as the inverse square root (ISR) $G^{-1/2}$.

If meanISR is not provided, return the 2-tuple $(X, G^{-1/2})$, otherwise return only matrix $X$.

If optional keyword argument transpose is true (default), $X$ holds the $k$ vectorized tangent vectors in its rows, otherwise they are arranged in its columns. The dimension of the rows in the former case and of the columns is the latter case is $n(n+1)/2$, where $n$ is the size of the matrices in 𝐏 (see vecP ).

if optional keyword argument ⏩ if true (default), the computation of the mean (if this is obtained with an iterative algorithm, e.g., using the Fisher metric) and the projection on the tangent space are multi-threaded. Multi-threading is automatically disabled if the number of threads Julia is instructed to use is $<2$ or $<3k$.

Examples:

using PosDefManifoldML

# generate four random symmetric positive definite 3x3 matrices
Pset = randP(3, 4)

# project and vectorize in the tangent space
X, G⁻½ = tsMap(Fisher, Pset)

# X is a 4x6 matrix, where 6 is the size of the
# vectorized tangent vectors (n=3, n*(n+1)/2=6)

# If repeated calls have to be done, faster computations are obtained
# providing the inverse square root of the matrices in Pset, e.g.,
X1 = tsMap(Fisher, ℍVector(Pset[1:2]); meanISR =G⁻½)
X2 = tsMap(Fisher, ℍVector(Pset[3:4]); meanISR =G⁻½)

See: the ℍVector type.

source

PosDefManifoldML.gen2ClassData — Function.

function gen2ClassData(n        ::  Int,
                       k1train  ::  Int,
                       k2train  ::  Int,
                       k1test   ::  Int,
                       k2test   ::  Int,
                       separation :: Real = 0.1)

Generate a training set of k1train+k2train and a test set of k1test+k2test symmetric positive definite matrices. All matrices have size $n$x$n$.

The training and test sets can be used to train and test any MLmodel.

separation is a coefficient determining how well the two classs are separable; the higher it is, the more separable the two classes are. It must be in [0, 1] and typically a value of 0.5 already determines complete separation.

Return a 4-tuple with

an ℍVector holding the k1train+k2train matrices in the training set,

an ℍVector holding the k1test+k2test matrices in the test set,

a vector holding the k1train+k2train labels (integers) corresponding to the matrices of the training set,

a vector holding the k1test+k2test labels corresponding to the matrices of the test set ($1$ for class 1 and $2$ for class 2).

Examples

using PosDefManifoldML

PTr, PTe, yTr, yTe=gen2ClassData(10, 30, 40, 60, 80, 0.25)

# PTr=training set: 30 matrices for class 1 and 40 matrices for class 2
# PTe=testing set: 60 matrices for class 1 and 80 matrices for class 2
# all matrices are 10x10
# yTr=a vector of 70 labels for the training set
# yTe=a vector of 140 labels for the testing set

source

PosDefManifoldML.predictErr — Function.

function predictErr(yTrue::IntVector, yPred::IntVector;
	          digits::Int=3))

Return the percent prediction error given a vector of true labels and a vector of predicted labels.

The order of arguments does not matter.

The error is rounded to the number of optional keyword argument digits, 3 by default.

See predict

Examples

using PosDefManifoldML
predictErr([1, 1, 2, 2], [1, 1, 1, 2])
# return: 25.0

source